+++ flink-libraries/flink-gelly/src/test/java/org/apache/flink/graph/generator/PathGraphTest.java
@@ -46,7 +46,7 @@ public class PathGraphTest extends GraphGeneratorTestBase {
 
 	@Test
 	public void testGraphMetrics() throws Exception {
-		int vertexCount = 100;
+		long vertexCount = 100;
 
 		Graph<LongValue, NullValue, NullValue> graph = new PathGraph(env, vertexCount)
 			.generate();
